stereo AY

+++ app/src/main/cpp/zxCommon.h
@@ -267,10 +267,10 @@ inline void wm8(uint8_t* address, uint8_t val) {
 }
 
 // возвращает количество миллисекунд
-inline long long currentTimeMillis() {
+inline u_long currentTimeMillis() {
     struct timeval tv;
     gettimeofday(&tv, nullptr);
-    return ((tv.tv_sec * 1000) + (tv.tv_usec / 1000));
+    return (u_long)(((tv.tv_sec * 1000) + (tv.tv_usec / 1000)));
 }
 
 inline uint8_t* ssh_memset(void* ptr, int set, size_t count) {
